semantic-search-api/pyproject.toml
[tool.poetry]
name = "semantic-search-api"
version = "0.1.0"
description = ""
authors = ["Debabrata Mondal <debabrata.js@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
uvicorn = "^0.17.6"
sentence-transformers = "^2.2.0"
SQLAlchemy = "^1.4.37"
pgvector = "^0.1.6"
psycopg2-binary = "^2.9.3"
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
